backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-rest-framework-mongoengine==3.4.1
mongoengine==0.28.2
pymongo==4.6.3
django-cors-headers==4.3.1
python-decouple==3.8
dnspython==2.6.1
